Move checkInstantiable to ConstructorConstructor
